Revert "Api agenda e ajustes no cadastro de usuario (#13)"
This reverts commit 14c7a48001df4751ae0de0743cf9368eb4976d04.
