+++ ima-api/pom.xml
@@ -28,6 +28,8 @@
     </scm>
     <properties>
         <java.version>21</java.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     </properties>
     <dependencies>
         <dependency>
@@ -69,7 +71,6 @@
             <artifactId>spring-boot-starter-mail</artifactId>
             <version>3.4.3</version>
         </dependency>
-
     </dependencies>
 
     <build>
