+++ ima-api/src/main/java/org/com/imaapi/model/Consulta/input/ConsultaInput.java
@@ -1,4 +1,4 @@
-package org.com.imaapi.model.consulta.input;
+package org.com.imaapi.model.Consulta.input;
 
 import lombok.Data;
 import org.com.imaapi.model.enums.ModalidadeConsulta;
