+++ ima-api/src/main/java/org/com/imaapi/model/Consulta/output/ConsultaOutput.java
@@ -1,8 +1,8 @@
-package org.com.imaapi.model.consulta.output;
+package org.com.imaapi.model.Consulta.output;
 
 import lombok.Data;
-import org.com.imaapi.model.usuario.Especialidade;
-import org.com.imaapi.model.usuario.Usuario;
+import org.com.imaapi.model.Usuario.Especialidade;
+import org.com.imaapi.model.Usuario.Usuario;
 import org.com.imaapi.model.enums.ModalidadeConsulta;
 import org.com.imaapi.model.enums.StatusConsulta;
