+++ ima-api/src/main/java/org/com/imaapi/model/Usuario/output/UsuarioOutput.java
@@ -1,4 +1,4 @@
-package org.com.imaapi.model.usuario.output;
+package org.com.imaapi.model.Usuario.output;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
