+++ ima-api/src/main/java/org/com/imaapi/repository/ConsultaRepository.java
@@ -1,6 +1,6 @@
 package org.com.imaapi.repository;
 
-import org.com.imaapi.model.consulta.Consulta;
+import org.com.imaapi.model.Consulta.Consulta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
