+++ ima-api/src/main/java/org/com/imaapi/repository/EspecialidadeRepository.java
@@ -1,6 +1,6 @@
 package org.com.imaapi.repository;
 
-import org.com.imaapi.model.usuario.Especialidade;
+import org.com.imaapi.model.Usuario.Especialidade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EspecialidadeRepository extends JpaRepository<Especialidade, Integer> {
