+++ ima-api/src/main/java/org/com/imaapi/repository/UsuarioRepository.java
@@ -1,6 +1,6 @@
 package org.com.imaapi.repository;
 
-import org.com.imaapi.model.usuario.Usuario;
+import org.com.imaapi.model.Usuario.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
