+++ ima-api/src/main/resources/application.properties
@@ -1,6 +1,8 @@
+appliction
+
 spring.application.name=ima-api
 
- ConfiguraÃ§Ã£o do banco de dados H2
+ Configuração do banco de dados H2
  spring.datasource.url=jdbc:h2:mem:ima-banco
  spring.datasource.driverClassName=org.h2.Driver
  spring.datasource.username=ima
@@ -10,8 +12,8 @@ spring.application.name=ima-api
  spring.h2.console.enabled=true
  spring.h2.console.path=/h2-console
  spring.jpa.hibernate.ddl-auto=create
- 
-# ConfiguraÃ§Ã£o do banco de dados MySQL
+
+# Configuração do banco de dados MySQL
 #spring.datasource.url=${MEU_BD}
 #spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
 #spring.datasource.username=${USER_BD}
@@ -20,7 +22,7 @@ spring.application.name=ima-api
 #spring.jpa.hibernate.ddl-auto=update
 #spring.jpa.properties.hibernate.globally_quoted_identifiers=true
 
-# ConfiguraÃ§Ã£o do Envio de email
+# Configuração do Envio de email
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.username=${EMAIL}
